app/core/config.py
from dotenv import load_dotenv
import os

load_dotenv()

class Settings:
	APP_NAME: str = os.getenv("APP_NAME","FastAPI APP")
	APP_ENV: str = os.getenv("APP_ENV", "dev")
	DB_HOST: str = os.getenv("DB_HOST", "localhost")
	DB_PORT: int = int(os.getenv("DB_PORT", 3306))
	DB_NAME: str = os.getenv("DB_NAME", "fastapi_db")
	DB_USER: str = os.getenv("DB_USER", "root")
	DB_PASSWORD: str = os.getenv("DB_PASSWORD", "sandeep")

settings = Settings()