requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
pymysql
python-multipart